+++ custom_components/climate_report/const.py
@@ -1,0 +1,3 @@
+DOMAIN = "climate_report"
+EVENT_REPORT_GENERATED = "climate_report_generated"
+SIGNAL_UPDATE = "climate_report_update"